pyproject.toml
[project]
name = "schedule-reminder"
version = "0.1.0"
description = "An autonomous, long-running **Reminder Assistant** powered by OpenAI and Resonate's Distributed Async Await."
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "openai>=1.82.0",
    "resonate-sdk>=0.5.3",
]
